upcoming: [DI-26185] - Added method for aclp enable check in contextual view (#12517)
* upcoming: [DI-26185] - Added method for  aclp enable check in contextual view

* upcoming: [DI-26185] - Added code comments

* upcoming: [DI-26185] - Added test cases

* upcoming: [DI-26185] - Updated logic

* upcoming: [DI-26185] - Renamed property from serviceType to capability

* added changetset

* upcoming: [DI-26185] - Updated test cases and renamed useIsAclpSupportedRegion

* upcoming: [DI-26185] - Replaced isAclpSupportedRegion method with useIsAclpSupportedRegion hook

* upcoming: [DI-26185] - Remove isAclpSupportedRegion file

* Updated changeset

* Removed unused imports

* updated changeset

* updated test cases

* updated test cases

* test: [DI-26185] - Updated test cases

* upcoming: [DI-26185] - Updated test cases


* updated test case

* updated test cases

---------
