+++ packages/manager/.changeset/pr-12517-added-1752583847765.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Added
+---
+
+ACLP: `useIsAclpSupportedRegion` method in `cloudpulse/Util/utils.ts file` ([#12517](https://github.com/linode/manager/pull/12517))
