+++ packages/utilities/.changeset/pr-12517-removed-1752767841927.md
@@ -1,0 +1,5 @@
+---
+"@linode/utilities": Removed
+---
+
+`isAclpSupportedRegion` helper method ([#12517](https://github.com/linode/manager/pull/12517))
